template_upload/requirements.txt
aioboto3==14.3.0
aiofiles==24.1.0
python-dotenv==1.1.0
